factored out mebus and intertechno devices
