+++ include/rtl_433_devices.h
@@ -12,5 +12,7 @@ extern r_device lacrossetx;
 extern r_device acurite5n1;
 extern r_device acurite_rain_gauge;
 extern r_device oregon_scientific;
+extern r_device mebus433;
+extern r_device intertechno;
 
 #endif /* INCLUDE_RTL_433_DEVICES_H_ */
