+++ src/CMakeLists.txt
@@ -82,7 +82,9 @@ add_executable(rtl_433
 	devices/elv.c
 	devices/lacrosse.c
 	devices/acurite.c
-	devices/oregon_scientific)
+	devices/oregon_scientific.c
+	devices/mebus.c
+	devices/intertechno.c)
 	
 set(INSTALL_TARGETS rtlsdr_shared rtlsdr_static rtl_sdr rtl_tcp rtl_test rtl_fm rtl_eeprom rtl_adsb rtl_433)
